+++ drivers/net/ethernet/meta/fbnic/fbnic_fw.c
@@ -163,7 +163,7 @@ static void fbnic_mbx_unmap_and_free_msg(struct fbnic_dev *fbd, int mbx_idx,
 	dma_unmap_single(fbd->dev, mbx->buf_info[desc_idx].addr,
 			 PAGE_SIZE, direction);
 
-	free_page((unsigned long)mbx->buf_info[desc_idx].msg);
+	free_page((uintptr_t)mbx->buf_info[desc_idx].msg);
 	mbx->buf_info[desc_idx].msg = NULL;
 }
 
@@ -214,7 +214,7 @@ static int fbnic_mbx_alloc_rx_msgs(struct fbnic_dev *fbd)
 		err = fbnic_mbx_map_msg(fbd, FBNIC_IPC_MBX_RX_IDX, msg,
 					FBNIC_RX_PAGE_SIZE, 0);
 		if (err)
-			free_page((unsigned long)msg);
+			free_page((uintptr_t)msg);
 	}
 
 	return err;
@@ -404,7 +404,7 @@ int fbnic_fw_xmit_test_msg(struct fbnic_dev *fbd,
 
 	err = fbnic_mbx_map_req_w_cmpl(fbd, test_msg, cmpl);
 	if (err)
-		free_page((unsigned long)test_msg);
+		free_page((uintptr_t)test_msg);
 
 	return err;
 }
@@ -441,7 +441,7 @@ static int fbnic_fw_xmit_simple_msg(struct fbnic_dev *fbd, u32 msg_type)
 
 	err = fbnic_mbx_map_tlv_msg(fbd, msg);
 	if (err)
-		free_page((unsigned long)msg);
+		free_page((uintptr_t)msg);
 
 	return err;
 }
@@ -548,7 +548,7 @@ int fbnic_fw_xmit_ownership_msg(struct fbnic_dev *fbd, bool take_ownership)
 	return err;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -766,7 +766,7 @@ static int fbnic_fw_xmit_heartbeat_message(struct fbnic_dev *fbd)
 	return err;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -870,7 +870,7 @@ int fbnic_fw_xmit_coredump_info_msg(struct fbnic_dev *fbd,
 	return 0;
 
 free_msg:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -959,7 +959,7 @@ int fbnic_fw_xmit_coredump_read_msg(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1077,7 +1077,7 @@ int fbnic_fw_xmit_fw_start_upgrade(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1153,7 +1153,7 @@ int fbnic_fw_xmit_fw_write_chunk(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1278,7 +1278,7 @@ int fbnic_fw_xmit_qsfp_read_msg(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1388,7 +1388,7 @@ int fbnic_fw_xmit_tsene_read_msg(struct fbnic_dev *fbd,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1594,7 +1594,7 @@ int fbnic_fw_xmit_send_logs(struct fbnic_dev *fbd, bool enable,
 	return 0;
 
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
 
@@ -1997,7 +1997,7 @@ int fbnic_fw_xmit_rpc_macda_sync(struct fbnic_dev *fbd)
 free_message_nospc:
 	err = -ENOSPC;
 free_message:
-	free_page((unsigned long)msg);
+	free_page((uintptr_t)msg);
 	return err;
 }
